chore(typescript): emit types for each file
