+++ lib/index.js
@@ -190,6 +190,11 @@ function jsonschema2md(schema, {
   return output;
 }
 
+/**
+ * Main function used in the CLI.
+ * @param {{ [key:string]: unknown }} args CLI arguments from user input
+ * @returns The generated Markdown files to the specified output directory
+ */
 async function main(args) {
   // parse/process command line arguments
   const { argv } = yargs(args)
@@ -310,9 +315,7 @@ async function main(args) {
     includeReadme,
     links,
     i18n,
-    // @ts-ignore
     language,
-    // @ts-ignore
     exampleFormat,
     includeProperties,
     header,
