+++ tsconfig.json
@@ -1,12 +1,13 @@
 {
   "compilerOptions": {
     "target": "es6",
+    "module": "commonjs",
     "moduleResolution": "node",
     "allowJs": true,
     "declaration": true,
     "emitDeclarationOnly": true,
     "removeComments": false,
-    "outFile": "./types/index.d.ts"
+    "outDir": "types"
   },
   "include": ["lib/index.js"],
   "exclude": ["**/locales/**/*"]
