+++ types/formatInfo.d.ts
@@ -1,0 +1,19 @@
+export function formatmeta(schema: any): {
+    longcomment: import("unist").Node;
+    shortcomment: any;
+    comment: string;
+    longdescription: import("unist").Node;
+    shortdescription: any;
+    description: any;
+    abstract: boolean;
+    extensible: boolean;
+    status: any;
+    identifiable: string;
+    custom: boolean;
+    additional: boolean;
+    definedin: {
+        text: any;
+        link: any;
+    };
+    restrictions: string;
+};
