+++ types/formattingTools.d.ts
@@ -1,0 +1,2 @@
+export function gentitle(titles: any, type: any): any;
+export function gendescription(schema: any): any;
