+++ types/keywords.d.ts
@@ -1,0 +1,2 @@
+export function keyword(str: any): any;
+export function report(): Set<any>;
